Reformat code and give option to bind to interface
We want to be able to bind our incoming socket to a specified address so
that we can be sure we are only responding to DHCP/BOOTP requests that
we intend (and not taking over these services for all our neighbors).
When specified, we only listen to this interface and we should only send
responses here as well.

Reformat the code to modified KNR used in the kernel.

Signed-off-by: Eric B Munson <munsoner@bu.edu>